Support TLS 1.2 TLS_RSA_WITH_AES_256_CBC_SHA256
Allow WebSocketClient to connect to Secure WebSocket server offering
cipher TLS_RSA_WITH_AES_256_CBC_SHA256.

The TLS 1.2 handshake is implemented with the help of external RSA service
which encrypts the client pre-master secret with the server's public
key. This assymetric encryption function is not available on the Roku
platform. After this step all other steps in this cipher suite involve
only symmteric encryption, which is natively available on the Roku
platform.

Implementation of the RSA service is availble in Python and C in the
files:
 -   util/crypto_server/rsa_server.py
 -   util/crypto_server/rsa_service.c

Most modern HTTPS servers will not allow handshake with this cipher but
implementing a newer cipher, like TLS 1.2 ECDHE-ECDSA-AES256-GCM-SHA384
or TLS 1.3 ciphers, requires third party service for every send or read
operation which renders the WebSocket functionallity useless, so this is
the best that can be done until Roku releases the roWebSocket component.